CDPS-1618 & CDPS-1620: Bring linting closer in line with HMPPS template (#243)
* Bring linting closer in line with template and add more type annotations

* Reduce sass deprecation warnings

* Only restart dev server once esbuild complete (hmpps-template-typescript#527)

Adding a simple debounce to the functions called from watchers prevent
spawning an excessive number of processes. This prevents issues with
running eth watchers on a large codebase, or when running several parent
processes in parallel, for instance, when running a development server
and integration tests at the same time.

---------

Co-authored-by: Fred Marecesche <froddd@users.noreply.github.com>